fix: accept float hrtime values in MonotonicTime
PHP 8+ returns float from hrtime(true); elapsed duration math now uses
float start markers and a narrowed intdiv operand.

Made-with: Cursor

+++ src/Runners/SequentialStreamRunner.php
@@ -173,7 +173,7 @@ class SequentialStreamRunner
         }
     }
 
-    protected function failStream(SwarmExecutionState $state, RunContext $context, int $contextTtl, Swarm $swarm, Throwable $exception, ?int $startedAt): SwarmStreamError
+    protected function failStream(SwarmExecutionState $state, RunContext $context, int $contextTtl, Swarm $swarm, Throwable $exception, ?float $startedAt): SwarmStreamError
     {
         $this->historyStore->fail($context->runId, $exception, $contextTtl);
         $this->contextStore->put($this->capture->terminalContext($context), $contextTtl);
